image/test/mochitest/test_discardAnimatedImage.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=686905
-->
<head>
  <title>Test that animated images can be discarded</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="imgutils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=686905">Mozilla Bug 686905</a>
<p id="display"></p>
<div id="content">
  <div id="container">
    <canvas id="canvas" width="100" height="100"></canvas>
    <img id="infinitepng" src="infinite-apng.png">
    <img id="infinitegif" src="animated1.gif">
    <img id="finitepng" src="restore-previous.png">
    <img id="finitegif" src="restore-previous.gif">
  </div>
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

/** Test for Bug 686905. **/
SimpleTest.waitForExplicitFinish();

var gNumDiscards = 0;

window.onload = function() {
  // Enable discarding for the test.
  SpecialPowers.pushPrefEnv({
    'set':[['image.mem.discardable',true]]
  }, runTest);
}

var gImgs = ['finitepng', 'finitegif', 'infinitepng', 'infinitegif'];

function runTest() {
  var animatedDiscardable =
    SpecialPowers.getBoolPref('image.mem.animated.discardable');
  if (!animatedDiscardable) {
    ok(true, "discarding of animated images is disabled, nothing to test")
    SimpleTest.finish();
    return;
  }

  // Draw the images to canvas to force them to be decoded.
  for (var i = 0; i < gImgs.length; i++) {
    drawCanvas(document.getElementById(gImgs[i]));
  }

  for (var i = 0; i < gImgs.length; i++) {
    addCallback(document.getElementById(gImgs[i]));
  }

  document.getElementById("container").style.display = "none";

  for (var i = 0; i < gImgs.length; i++) {
    requestDiscard(document.getElementById(gImgs[i]));
  }
}

function step2() {
  document.getElementById("container").style.display = "";

  // Draw the images to canvas to force them to be decoded again.
  for (var i = 0; i < gImgs.length; i++) {
    drawCanvas(document.getElementById(gImgs[i]));
  }

  SimpleTest.finish();
}

function addCallback(anImage) {
  var observer = new ImageDecoderObserverStub();
  observer.discard = function () {
    imgLoadingContent.removeObserver(scriptedObserver);
    gNumDiscards++;
    ok(true, "got image discard");
    if (gNumDiscards == gImgs.length) {
      step2();
    }
  }
  observer = SpecialPowers.wrapCallbackObject(observer);

  var scriptedObserver = SpecialPowers.Cc["@mozilla.org/image/tools;1"]
                           .getService(SpecialPowers.Ci.imgITools)
                           .createScriptedObserver(observer);

  var imgLoadingContent =
    SpecialPowers.wrap(anImage)
                 .QueryInterface(SpecialPowers.Ci.nsIImageLoadingContent);
  imgLoadingContent.addObserver(scriptedObserver);
}

function requestDiscard(anImage) {
  var request = SpecialPowers.wrap(anImage)
      .QueryInterface(SpecialPowers.Ci.nsIImageLoadingContent)
      .getRequest(SpecialPowers.Ci.nsIImageLoadingContent.CURRENT_REQUEST);
  setTimeout(() => request.requestDiscard(), 0);
}

function drawCanvas(anImage) {
  dump(anImage + "\n");
  var canvas = document.getElementById('canvas');
  var context = canvas.getContext('2d');

  context.drawImage(anImage, 0, 0);
  ok(true, "we got through the drawImage call without an exception being thrown");
}

</script>
</pre>
</body>
</html>
